move ConfigurationNode
